requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
bazel==0.0.0.20200723
beautifulsoup4==4.9.1
bert-base==0.0.7
bert-serving-client==1.10.0
bert-serving-server==1.10.0
Brotli==1.0.9
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Cython==0.29.14
distlib==0.3.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Cors==3.0.9
Flask-JSON==0.3.4
FlyAI==0.7.1
gast==0.2.2
gensim==3.8.3
gevent==20.6.2
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
GPUtil==1.4.0
greenlet==0.4.16
grpcio==1.31.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
install==1.3.4
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
JPype1==0.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
lxml==4.6.1
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.1
Pillow==7.2.0
prettytable==2.0.0
progressbar==2.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyhanlp==0.1.66
pymongo==3.11.0
PyMySQL==0.10.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
qrcode==6.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.20.4
scipy==1.5.4
six==1.15.0
smart-open==4.0.1
soupsieve==2.0.1
tensorboard==1.14.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-serving-api==1.14.0
termcolor==1.1.0
threadpoolctl==2.1.0
tqdm==4.48.2
treenode==0.1.5
ufile==3.2.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
